[refactor] Changes to namespace concept.
This implements the changes discussed earlier with regard to
namespaces and configurations. The namespace concept is removed in
favour for a more engine-driven approach, where the resolver plays a
greater role in where the configuration is being located on disk. The
simple base resolver implementation locates its configurations
grouped by engine in order to ensure uniqueness, however future
resolver implementations are free to create their own arbitrary
structures. By moving the logic of config path resolution to the
resolver, the configuration class could be simplified further and now
simply operates on a given configuration location without the ability
to compute this location by itself.