Adds automatic logging to file as part of engine bootstrap.
When an engine is started via the bootstrap_engine call, a file logger
is now automatically created for the sgtk log tree, writing to a
log file. This follows the pattern we already have in desktop and
means that logging to file automatically and consistently happens in
a single place in the code. Engines can later on re-route their
various loggers into the sgtk log namespace so that engine log
messages are also written to file. In addition to a file logging
facility, each bootstrap integration can add its own UI based logging
 - this is unrelated to this change.

 Closes #249