[minor] Improved logic around dev workflows and cloning.
Previously, a config with a git descriptor would always be cloned
whenever the bootstrap happened. The assumption was that if you use a
git descriptor, you want to clone it into place so that you can to
development from that location. This is not always the case,
sometimes you want the git descriptor to follow the same logic as
everything else. Therefore, a flag now controls this behaviour so
that it can be specifically turned on if needed.