[minor] Manual descriptors are now marked as immutable.
Previously, manual descriptors were incorrectly marked as immutable
by the system, highlighting the fact that these are maintained
manually by users and not automatically by the system. However, this
is conceptually incorrect - because a manual descriptor is versioned,
each entry in the cache really is unique and toolkit should be able
to safely assume that these items won't change after they have been
installed.