[minor] Fixes issues around bootstrap config rollback
The code would make a backup folder based on a second-resolution
timestamp. This was fine most of the time, but when something failed
and the code attempted a rollback, the rollback could happen within
that same second as the previous install, causing the rollback to
attempt to create the same backup folders as the original backup,
causing errors. This code now adds checks to ensure that a unique
backup folder is created.