for #35713: add deprecation warning to context_additional_entities hook
This is an old hook that never matured to a point for wide use and generates more questions than answers. Support for it is limited and there are now better ways to solve the challenges it was attempting to solve. Because there may be studios that still use it, we won't remove it until we can confirm all clients have removed it from their configs (somehow).