for #35739: modify shared core warning when upgrading core
The note emitted when upgrading core was originally written when Toolkit used a shared studio core by default. Since this is currently no longer the case and SG Desktop localizes your core by default when setting up a project, we've re-worded the note to make it clearer and a little less scary. This was previously causing some confusion for clients who didn't even necessarily know what a shared core was, which means this shouldn't affect them anyway. Added links to documentation as well.