[minor] log message update - deprecation of context_additional_entities
- For #35492: Updated note about core upgrade potentially affecting multiple projects.
The note emitted when upgrading core was originally written when Toolkit used a shared studio core by default. Since this is currently no longer the case and SG Desktop localizes your core by default when setting up a project, we've re-worded the note to make it clearer and a little less scary. This was previously causing some confusion for clients who didn't even necessarily know what a shared core was, which means this shouldn't affect them anyway. Added links to documentation as well.

- For #35713: Added a deprecation warning to the context_additional_entities hook.
This is an old hook that never matured to a point for wide use and generates more questions than answers. It doesn't do what clients think it does and there are now better ways to solve the challenges it was attempting to solve. Because there may be studios that still use it, we won't remove it until we can confirm all clients have removed it from their configs (somehow).

closes #260