#35511 Adds a new command to dump configs to stdout/file as full/sparse
These changes adds a new `dump_config` action to the `./tank` command. The
action allows the user to write out the contents of an existing environment to
`<stdout>` or a file. It supports both full and sparse config dumping.

The changes correspond to recent modifications to the `develop/0.18` branch of
core to allow for the use of sparse configurations (not requiring configuration
settings when the default value is used).

The usage for the `dump_config` option looks like this:

```
./tank dump_config env_name [--sparse | --full] [--no-debug-comments] [--file=/path/to/output/file.yml]
```

The `--sparse` and `--full` options are mutually exclusive. If neither are
included, then the config will be dumped as it exists on disk.

By default, the action writes to `<stdout>`. If the `--file` option is supplied
then it will write to the specified path.

The `--no-debug-comments` option will prevent additonal debug comments from
being included in the dumped config (the default behavior is to include them).
The debug comments identify which whether the settings in the dumped output
match the default value or not. They also include the default value when the
setting differs. Also included is the path to the manifest where the bundle's
default values originate.

The dumping is made possible via changes to the `WritableEnvironment` object in
tank.platform.environment. A new public method has been added: `dump()`.
This method accepts an open file handle to write to. The `WritableEnvironment`
itself is heavily tied to its location on disk, and this methods allow
dumping that source file to the handle provided.

The other change of note in the `WritableEnvironment` object is the new private
method called `__write_data_file()`. This logic is pulled directly from
`__write_data()` (which now calls `__write_data_file()`) and allows you to
write the supplied yaml data to a file handle. This is what facilitates the new
`dump` method. None of the logic has changed, only exposed internally as a
separate private method.

The `validate` action was also updated to include the ability to process a
subset of configs. You can now filter the environment names to process. Before
this change, the validate command would validate each environment file it
found. Now you can specify a list of environment names to process:

```
./tank validate asset_step shot_step
```

The above command would validate/dump each of the `asset_step` and `shot_step`
environments.

Closes #263