For #34433, framework update fix in includes #257
When adding missing frameworks to environment files, we never made sure to always add all dependencies within the file of the bundle requiring it. One tricky scenario wasn't covered.

Let's pretend we have an app A, having a dependency on framework B, all registered inside a common include. At some point in the future, app A is released again and framework B now has a dependency on framework C. If C had never been used in the environment, then the update code would have added the framework C to the include file, so that any environment that uses the app A would have access to framework B's and C's definition. The rule was: always add missing dependencies inside the environment file requiring it.

Here's an edge case however. Suppose that framework C is already installed in one environment, but not the others. The previous code would look for missing frameworks and determine that frameworks B and C are not missing when updating the first environment. If the user then didn't update other environments, then they would have been all be broken because these other environments don't have framework C registered in them.

With this fix, when looking for missing frameworks, we are always starting the search from the file where the bundle that requires it is registered. This means that in the above scenario, we will start looking from the include file itself. This means that the update code will then determine that framework C is not available and add it to the include file. Which means that any environment who is using that app will have access to the complete list of registered frameworks.

This branch also introduces a unit test framework to test app store updates. It provides a patcher class that allows to replace part of the AppStore code so that we can programatically define AppStore bundles and pretend we're downloading updates from the app store.

Closes https://github.com/shotgunsoftware/tk-core/pull/257