#34528 Now defaults to ruamel yaml parser
This switches the yaml parsers around so that the ruamel (the new,
smarter parser) is the default and you can "opt out" of this
if needed.

The previous tank command flag --preserve-yaml has been removed
and replaced by a --use-legacy-yaml flag instead that can be added
if you want to opt out. Similarly, the TK_PRESERVE_YAML env var
has been replaced by a TK_USE_LEGACY_YAML.

Closes #237.