#35786 Fixes full-sync path cache issues when using concurrent projects
This fixes an issue causing unnecessary full path cache syncs in
certain situations if you work on more than project at once.
Previously, the event log id "marker" used to check the highest event
log id synced in the local path cache was not limited by the current
project, causing it to store the highest event log id for all projects.

If more than one project was generating new paths in Shotgun, this
could result in the wrong id marker being stored in the local cache.
Once the wrong marker was in the local cache, the sync logic would
get confused and fall back on a full sync if the max id marker
record didn't belong to the current project.

Closes #266