#34751 Ensures host and http_proxy strings are not unicode.
In RV (and possibly elsewhere in the future) we need to be careful about not propagating unicode strings into the Python API when establishing the connection to Shotgun. The host and http_proxy strings are concatenated with data on upload, which can cause Python to try to convert the contents of a file to unicode which then throws a UnicodeDecodeError.

Closes #268.