#35516 merge followup: Removed unused logic.
Because the way the app store logic has been refactored in 0.18,
changes need to be manually merged. This backs out of some logic
which was specifically added solely for the purpose to be able to set
the app store connnection timeout. We did some adjustments in
5554d12b8e0dc6d4245afd85a1fbbafd0eb7c427 to make the app store
timeout changes 0.18 compatible, but we never removed the code which
was added in 0.17, and that is no longer used anywhere.