Backwards compatibility for auth config files (#277)
The desktop assumes that if you have authenticated on the site level, you are also authenticated on the project level. In 0.18, a new set of root paths (pending our discussion) are introduced and this means that there was a situation where auth info was stored on disk and readable for a 17 core but not for an 18 core. So if your project was 0.18 and your site config was 0.17, things wouldn't work.

This PR adds a fallback so that if the 0.18 config isn't found, it checks for the 0.17 one as well (similar to how the cache hook works). If none of them is found, the 0.18 path is returned.