#36597 Less destructive core swapping
The previous core swapping implementation simply removed all previous sgtk modules from sys.modules. This would mean that all weak references going back to the modules kept in sys.modules would be lost - including when classes had subclasses a component and some of the global state that was floating around.

This fix makes the behaviour a lot less brutal - instead of GCing the old module objects, they are moved in sys.modules (a guid prefix is attached) to keep the previous module objects in memory. For old code where the binding has already happened, everything will remain the same - for new code that is about to be executed, the new version of core will be used.

This has the consequence that any internal bindings within sgtk will be switched round, but any external bindings will be retained and will be using the old code. As with all code swapping/reload, there is always complexity, however I think this solution offers a better compromise between stability and cleanliness than the previous one. This has been tested in RV with the custom bootstrap, which previously failed.

Closes #284