[minor] Fixed bug in error detection based on log entries.
Previously, there was a primitive check counting the number of log
messages emitted, assuming that if there were any messages emitted,
this would equate to an error. With the 0.18 improved logging, there
is a lot more debug logging going on, so the logic would yield false
positives whenever debug was emitted. This commit adjust the log
detection so that it only triggers when log levels above warning are
reported back to the logger.