Version: latest token only injected into descriptor if not set
Previously, if you request a latest version for a descriptor, it
would always replace whatever version token you pass in with an
artificial version: latest token in the descriptor. Once that
descriptor was resolved, it would then proceed and request the latest
descriptor from that artificial descriptor. This doesn't work for
manual descriptors, where the latest method just returns "self".
Instead, now the version number is preserved when passed in, which
means latest for manual descriptors now function correctly.