#36909 entry point support for local cache manager
This is a subtle yet important change to allow for a fully sandboxed and consistent setup when running TaaP entry point name spaces. Previously, we only distinguished between toolkit configs that either had a pipeline config id or not. Most TaaP setups will not have a pc id in shotgun and this change ensures that a TaaP setup cache structure works the same way as cache structures for toolkit configs that have a pipeline config id. With this change, the entry point is used as part of the default folder key if a pipeline config id doesn't exist, ensuring consistency in the cache structure on disk.

Further changes:

- This change changes the interface of the cache hook. Created new interfaces with graceful fallbacks for any user that has overridden it. Any existing implementation should work perfectly as long as they don't try to run TaaP with it. A log warning is issued.

- Pipeline configuration class now tracks entry point alongside project id and pc id.

- Untangled a long standing issue where the core code tried to be backwards compatible by explicitly changing the site config (project id None) into project 0. This was causing the cache structure to be a bit messy. Now it is clean and backwards compatible and supports old hooks.

Closes #289