#30005 Support for Shotgun specific storage syntax
On Shotgun versions where supported (6.3 and later), the register publish call now explicitly sets the storage it wants to associate with a published file. If not supported or if the storage is not found in Shotgun, it gracefully falls back on the previous implementation, e.g. passing the entire path to Shotgun and relying on Shotgun to compute which storage to associate with. Also added debug logging around the publish process.

Closes #294