#37386 Fixed issue preventing project copy in Shotgun desktop
Fixed a bug which was causing certain items to be missing from a new project's app cache when it the project was created from another project's config. This issue was caused by the fact that toolkit maintains a global in-memory cache of descriptors. The assumption is that if a descriptor points at Core 1.2.3, this descriptor can be used by any process that wants to pick up v1.2.3.

Special logic was used inside the localize method to determine if an item should be copied or not, making assumptions about how descriptors were implemented under the hood. These assumptions not only conflict with this global cache, but generally going forward with the whole concept of the global bundle cache (so a good thing that we found this issue).

The most reliable fix I can think of here is to essentially copy the install folder across at localization time - we cannot make any clever assumptions anymore around descriptors and install locations since descriptors now can be stored in many different places etc. This means that the implementation is going back to something that is similar to its original design.

Notable points:

- Everything in the install folder is now copied across, including versions that are not used by the config. This is not necessarily a bad thing, but in some cases can increase the I/O burden.

- Modified the default permissions for the copy file I/O method. This was introduced in 0.18 and they were wrong.

- This change may affect larger setups which rely on complex localization workflows. Such clients are unlikely to be affected negatively by this change but should check carefully.

- Improved use of internal I/O methods

- Added logging around descriptors.

Closes #306