#37459 Adds exception logging to hooks
The hook interface is a boundary where toolkit executes user code. With 0.18 and its improved logging, we can ensure to consistently trap any errors passing through this boundary.

This change adds a low level exception logging trap for all hooks in toolkit, ensuring that whenever a hook fails to load or execute, this is logged to the logger (and the log file):

2016-07-15 09:43:48,642 [90897 ERROR sgtk.core.hook] Exception raised while executing hook '/mnt/wintermute/configs/18_mac/install/core/hooks/tank_init.py'
Traceback (most recent call last):
  File "/mnt/wintermute/configs/18_mac/install/core/python/tank/hook.py", line 506, in execute_hook_method
    ret_val = hook_method(**kwargs)
  File "/mnt/wintermute/configs/18_mac/install/core/hooks/tank_init.py", line 25, in execute
    a = b + c
NameError: global name 'b' is not defined

The exception itself is then bubbled up to the caller.

Note: Because the caller may implements its own exception trapping and reporting, it is possible, even likely, that the hook exception will be logged twice. But this specific message is useful because (apart from happening every single time) it provides a clear error message that only contains the callstack from the user's code (e.g the hook code).

Closes #307