#36879: Adds engine events.
Engine events are emitted by the engine when something occurs in the host DCC application that child sgtk apps need to be notified of. As of this initial implementation, the FileOpenEvent is available. When an event is emitted by an engine, each of its child apps will have their event_engine method called, regardless of the event type, and then the event-type-specific method will also be called (ie: event_file_open). This closely mirrors the Qt framework's event system.

Closes #302.