#37459 Relaxed hook exception logging to only cover core hooks.
The implementation added via a4ec2724832e685165080d137f9545904bd67a5d
was affecting cases where an app and a hook had established a mutual
agreement that exceptions should travel from the hook to the
application runtime - in these cases, a false positive would be
generated in the form of a logged exception which wasn't actually an
error at all. The logging is therefore now restricted to only cover
core hooks.