#37448 Fleshed out dump_config command usage
This change is limited to the dump_config command and fleshes out the usage and
information displayed when invalid/insufficient parameters are passed. Also
includes a little more output to indicate when the process completes.

The "--no-debug-comments" option has been changed to "--no_debug_comments" to
be consistent with other commands.

Closes: #315.