#37633 Localizes apps correctly for 0.17 cores
Fixes an issue with localization when a 0.18 core is running in desktop and is used to run the project setup for a project which is using a 0.17 core.

If you go through the setup wizard in desktop and pick an old 0.17 project, the project is copied across and the localize is trying to be smart about enumerating the descriptors and only copy bundles that are being used across to the new project. However, this bundle copying logic is based on the new 0.18 bundle disk structure, meaning that as soon as the new project starts up with its 0.17 core, it will not find the localized bundles because they reside in a location which is unknown to the 0.17.

This fix effectively implements the approach first done (and CRed) in #306. If a 0.17 based project is detected, the brute force copy kicks in. For 0.18 projects, the smart copy logic is used.

Closes #320