#38064 .git not copied across at project setup
Fixes a regression where projects set up with a git based config didn't end up with a .git folder.

This was an oversight when trying to improve the behaviour of the copy_folder generic I/O method. With this change, the descriptor.copy() command has been updated to always include .git folders when a git descriptor (tag or branch) is copied across. Also fixes an issue with how the skip list is handled within the copy_folder() method.

Closes #335