#38304 #37933 entry point > plugin id, site level overrides
This renames entry_point to plugin_id across the code. Existing plugins are likely (but not guaranteed) to be backwards compatible. 

Also implements site level overrides for pipeline configurations, allowing for the setup of a single pipeline configuration that is used by a plugin (or plugins) across all projects for a site. For more info, see #337.

Closes #337