#29229 Context from path returns step with tasks
This resolves an old issue where it was possible to do the following: If you had a folder schema using task folders but not steps (e.g. project/sequence/shot/task) and you did context_from_path() on a folder, the context returned would have the task set but not the step.

This is unintuitive since the task implies that we know the step. This change fixes this issue via the folder creation, so all new folders that created will have step information stored alongside task information in the path cache (as a secondary path cache entry).

Also split up the folder types file into an entire submodule of things. 

Closes #343