#39255: Only log metrics for custom core hooks once per session
These changes ensure custom core hook and custom core hook methods have metrics
logged only once during a user's session. The `PipelineConfiguration` now has
a class member that keeps track of the already logged hook and hook.method
names.

The original plan was to add an option to the `log_metric()` calls to tell the
system to only log this metric once. This would require one of 2 approaches.
Either the metrics system knows how to identify the metric to be logged by
hashing it somehow (or introspecting the data which varies depending on the
type of metric) or the `log_metric()` methods would need to be supplied with a
key for identifying the metric. Either of these is doable, but a bit more work
than this small change. Maybe this is sufficient for now and if we find we need
to repeat this somewhere else in the code, then we revisit? We can discuss.