#38647 Adds a method to load newly bundled core fonts in Qt
This introduces a new protected method in the Engine class called
_ensure_core_fonts_loaded. This method loads any fonts that are bundled with
core. Accordingly, this change also bundles Open Sans fonts with core.

Fonts are loaded automatically during the engine's __init__ if a QApplication
is available. Similarly fonts are loaded if a subclass uses
_initialize_dark_look_and_feel.

Open Sans is now the default font when using the bundled dark look and feel.

TankQDialog has been updated to use Open Sans for its header and details page.
It does not enforce Open Sans on the embedded widget section for fear of
forcing our client's hand when it comes to specifying a default font.

For apps that wish to make use of the bundled fonts, they would only need to
add something like the following in their style.qss file:

QWidget {
    font-family: "Open Sans";
}

Closes #355.