#38647 Fixes issue where busy dialog text is cut off
This change simply makes the busy dialog slightly larger. This seemed
preferable to making the font size smaller or trying to do anything dynamic.
I searched through all existing calls to this in our repos, and the given
case seems to be the longest message displayed. This will give us a little more
room to work with and we can do something a little smarter in the future if
need be.