For #38957 and #38959, Adds new method platform.create_engine_launcher()
Adds a new factory method to platform called `create_engine_launcher` that a generic toolkit launch app can call to generate a SoftwareLauncher object for a particular engine. The SoftwareLauncher base class is implemented in platform, but the methods expected to be called by a launch app (`scan_software()` and `prepare_launch()`) are abstract and must be implemented by engine specific subclasses. The `scan_software()` method implements DCC-specific logic to scan a filesystem for software installations and returns a list of SoftwareVersion objects, which are merely  property bags containing information about the specific version of the DCC to launch. This information is wrapped in a class rather than using a dict to make it easier to extend in the future. The `prepare_launch()` method returns a LaunchInfo object containing all the pieces to carry out the launch - the path to execute, the args to pass, environment variables to set. Both the SoftwareVersion and LaunchInfo classes are implemented in platform. The Bundle and Engine classes were slightly modified in order to share code with the new SoftwareLauncher class and create_engine_launcher method.