#40101: Add a caching policy parameter to bootstrap. (#375)
Add property "ToolkitManager.caching_policy" to specify the config caching policy to use when bootstrapping.

Setting it to "ToolkitManager.CACHE_SPARSE" will make the manager download and cache the sole config dependencies needed to run the engine being started. This is the default caching policy.

Setting it to "ToolkitManager.CACHE_FULL" will make the manager download and cache all the config dependencies.