For #40080, download_url(): Updated SG auth/added new param (w/default)
Integrated authorization features from the Shotgun API download_attachment() method that will allow input URLs containing a shotgun server name to correctly redirect to an s3 location. 

Added `use_url_extension` parameter to append the filetype extension of the resolved url to the input download location. In some cases, the target content of the url is not known beforehand. For example, the url "https://my-site.shotgunstudio.com/thumbnail/full/Asset/1227" may redirect into "https://some-site/path/to/a/thumbnail.png". In such cases, setting the `use_url_extension` parameter to True will cause the method to append the file extension of the resolved url to the filename passed in via the location parameter. So for the urls given above, you would get the following results:

    - location="/path/to/file" and use_url_extension=False would return "/path/to/file"
    - location="/path/to/file" and use_url_extension=True would return "/path/to/file.png"