System settings are now read by the SystemSettings class.
The SystemSettings is now responsible for reading system wide settings, just like UserSettings is used to read the current user settings. This is being used in the defaults manager, which uses it as a fallback if shotgun.yml and the user settings have not specified an http proxy.
Unit tests have also been refactored to take into account the change of behaviour of UserSettings.