#39913 : Fixes tag test git repo fixture. (#400)
We're now using a bare repo that contains the tags that we need for the tag sorting unit tests. This works without the need for a submodule, and also fixes issues with the previous attempt at building this fixture that resulted in a missing .git directory, which broke the tests.