#41090 Fixes to hook.disk_location (#407)
Fixes a bug which would only happen when you have complex inheritance chains for hooks. We now commonly apply the following sort of pattern:

some_hook: {self}/collector.py:{self}/maya.basic/collector.py

If the disk_location property is called from within {self}/collector.py, we expect the path returned to be {self} and when we call it from within {self}/maya.basic/collector.py, we expect it to be {self}/maya.basic. This was not previously the case - in the past, it would always return the path to the leaf class, e.g. in this case the path {self}/maya.basic.

This fix addresses this. In the case disk_location is called from outside a hook, we always return the disk location of self, eg. the location of the leaf hook in the inheritance.