Makes `products` and `versions` first class citizens of `SoftwareLauncher`
The SoftwareLauncher can now accept a list of versions and products. This makes
it possible for the launcher to know if a SoftwareVersion is supported. New
`products` and `versions` properties added to the object. Also, the
`is_supported_version` method now accepts a `SoftwareVersion` object instead
of a string. The method can now check the supplied object's minimum version and
can also check to see if it's product and version are within the specified
lists.