For #40914, branches and revisited releases
Added branch property to IODescriptorAppStore. Allowed it to be specified in descriptor dict and uri. Added unit tests for it.
Ensured that if a branch is set for an IODescriptorAppStore, only versions for this branch are considered. Branch versions are ignored as well for "master" versions.
Revisited _find_latest_tag_by_pattern to allow different version conventions than the strict vdigit.digit.digit currently enforced, and leveraged distutils.LooseVersion.
Added branch accessor on Descriptor.
Started to add unit tests for new features.