For #40914, revisited patterns
Allowed pattern to not always start with a "v".
Enforced patterns using the "v" pattern to only match versions with a leading "v".
Enforced number of tokens in the pattern to be matched by a version number.
Convert LooseVersion to a string before returning it.
Added unit tests for the various version patterns we might want to support: regular vx.x.x, topic releases with a leading ticket number, releases with a leading software release.