#41300 improved publish resolve (#425)
Improves the Hook.get_publish_path() logic to handle smarter resolves and better error reporting. 

The method will attempt to resolve the publish data into a path by applying several recipes:

First, the resolve_publish core hook will be called. This hook can be used either in order to override any built-in behavior or to implement support for url schemes, uploaded files or other modes which the default implementation currently doesn’t support. The hook returns None by default, indicating that no resolve overrides are present. If no path was resolved by the hook, the method will check if the publish is associated with a local file link and if so attempt to resolve this to a path (for more details see below). Next, it will check if the publish is linked to a file:// url and attempt to resolve this. Lastly, if the publish still cannot be resolved, a PublishPathNotSupported is raised.

If the publish is a local file link, its local OS representation will be used. Local file links are generated automatically by register_publish() if the path you are publishing matches any local file storage defined in the Shotgun Site Preferences. Read more about local file links here.

Local file storage mappings in Shotgun are global and affect all machines, all users and all projects. Sometimes there are cases when you want to override these globals to behave differently on a specific machine or for a user. This can be done by defining an environment variable named of the form SHOTGUN_PATH_OS_STORAGENAME. For example, if you are running Windows, your PRIMARY storage has been set up in Shotgun to be F:\ and you want paths on your machine to resolve to G:\ instead, set an environment variable SHOTGUN_PATH_WINDOWS_PRIMARY=G:\. The corresponding variables on Linux and Mac would be SHOTGUN_PATH_LINUX_PRIMARY and SHOTGUN_PATH_MAC_PRIMARY.

The method also supports the resolution of file:// urls. Such urls are not multi platform and the local storages and environment variables will therefore be used to try to resolve such paths in case of ambiguity. First, local storage settings will be downloaded from Shotgun and added to a path translation table. Next, similar to the local file linking above, any environment variables of the form SHOTGUN_PATH_WINDOWS|MAC|LINUX or SHOTGUN_PATH_WINDOWS|MAC|LINUX_STORAGENAME will be added to the translation table.
The file:// path will be compared against all the existing roots for all operating systems. The first match detected will be used to translate the path into the current operating system platform.
If no root matches, the file path will be returned as is.

For more information, see the API docs. In addition to the above:

- Added a general utility method sgtk.util.resolve_publish_path(tk, sg_publish_data). This is what is being called by Hook.get_publish_path()
- Hook.get_publish_paths() has been marked as deprecated and taken out of the docs
- Improved sphinx docs for Hook.get_publish_path()
- Improved sphinx docs for exceptions in general
- Added new exceptions for publish lookups
- Added new core hook resolve_publish() for business logic implementation of custom resolvers.
- Added Hook.sgtk and Hook.tank properties for easy access to core API instance from hooks
- Fixes bug with resolution of file:// urls with windows drive letters
- Split the shotgun.py file into multiple smaller chunks
- Added in-memory caching of local storages for performance
- Added setters to ShotgunPath os accessors .macosx, .windows, .linux and .current_os
- Added ShotgunPath.normalize() method for path sanitation
- Added extensive test coverage of the changes above.