#42005 App store descriptor label support. (#440)
Adds support for an optional label parameter for app store descriptors.
This allows for a workflow where a single stream of updates from the app store can target several different plugin releases. The primary use case for this is that you want a single config to drive plugin based integration. For more info, see the PR.