#41300 Storage followup (#445)
QA fixes for #41300 - handling of how storages are resolves when a publish is resolves into a path. Tweaks to sphinx docs with better examples and clearer docs. Added support for adding an os platform to a local storage at runtime via an env var so that you can augment a local storage that is missing an os platform at runtime. this brings symmetry to the operations you can perform across file:// links and local storage publishes.