Fixes a bug in the bootstrap API’s resolver.
The desired behavior to use the entity’s relevant platform-specific path if it’s set, followed by the “descriptor” field’s contents, if set, followed by the legacy “sg_descriptor” field’s contents. If no usable path is found after that order of lookup, an exception is raised.

Previous to this, if a PC entity didn’t have a path set for your operating system, it would raise an exception even if there was a usable descriptor set that would have been resolved into the correct path on disk.