For #43024: Handles an old core with a version of “HEAD” in its info.yml. (#467)
In this case, we don't know if the target tk-core is new or old. We attempt to treat it as "new" and use the modern code path for bootstrapping the Shotgun engine. If that fails, then we also try to fall back to the legacy pathway. This should be exceedingly rare, as it would require a locked-off site config that contains a tk-core that's been cloned from Github, rather than being pulled from the Toolkit app store.