For #43060: Refactors the legacy path for tk-shotgun bootstrapping. (#481)
This resolves an issue where the wrong entity type was passed to start_shotgun_engine when the legacy start was required for the shotgun engine.

This removes legacy compatibility logic from sgtk.platform.engine and puts that responsibility on the bootstrap manager instead. That approach also allows us to remove the core version check that was required in the previous implementation.