For #42873: [minor] Removes a log_exception and re-raise.
The reason this has been removed is that in some cases, most notably in the bootstrap API, it is reasonable and likely to trap the engine init exception and handle it gracefully at a higher level. In that case, you definitely don’t want the user to see the nasty exception in logs when it doesn’t correspond to any actual problem. The prime example of this is that when executing an sgtk command via the tank action menu in the web app, exception logs like this bubble back up to the client and are shown to the user after command execution. Even if that execution was successful, they get a heinous Python error thrown in their face.